requirements.txt
aiohttp==3.5.4
aiopg==0.16.0
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
iniherit==0.3.9
jsonschema==3.0.1
multidict==4.5.2
psycopg2-binary==2.7.7
pyrsistent==0.14.11
PyYAML==3.13
six==1.12.0
SQLAlchemy==1.3.1
text-unidecode==1.2
typing-extensions==3.7.2
yarl==1.3.0
